Remove global namespaces (process.namespaces assignment)
Adding properties to process (possibly some other builtin node modules
too?) creates issues in ESM projects that use this library and use jest
for running tests.

The relevant jest issue: https://github.com/facebook/jest/issues/12650

The problem is specifically with CommonJS modules that have side effects
like this and are used in ESM projects, if the description is correct.

The jest error that was caused by this:

    ReferenceError: Export 'namespaces' is not defined in module
        at /lune/backend/node_modules/jest-runtime/build/index.js:1869:16
        at Array.forEach (<anonymous>)
        at ModuleWrap.module.context.context (/lune/backend/node_modules/jest-runtime/build/index.js:1867:34)
        at SyntheticModule.evaluate (node:internal/vm/module:226:23)
        at evaluateSyntheticModule (/lune/backend/node_modules/jest-runtime/build/index.js:2617:16)
        at ModuleWrap.<anonymous> (node:internal/vm/module:316:24)

Internal Lune ticket: https://linear.app/lune/issue/LUN-1889/fix-the-referenceerror-export-namespaces-is-not-defined-in-module-jest